Logger refactoring (#435)
* rename Logger to GlobalLogger, BaseLogger to Logger. Added new registerContextAwareLoggerToContainer function.

* Fix issues with trying to mock abstract logger class
