+++ packages/logger/src/base-console-logger-client.spec.ts
@@ -7,8 +7,8 @@ import { IMock, Mock, MockBehavior, Times } from 'typemoq';
 import * as util from 'util';
 
 import { BaseConsoleLoggerClient } from './base-console-logger-client';
-import { LogLevel } from './base-logger';
 import { BaseTelemetryProperties } from './base-telemetry-properties';
+import { LogLevel } from './logger';
 import { ScanTaskStartedMeasurements } from './logger-event-measurements';
 import { LoggerProperties } from './logger-properties';
