+++ packages/logger/src/global-logger.ts
@@ -2,12 +2,17 @@
 // Licensed under the MIT License.
 
 import { injectable } from 'inversify';
+import { BaseTelemetryProperties } from './base-telemetry-properties';
 import { Logger } from './logger';
 import { LoggerClient } from './logger-client';
 
 @injectable()
-export class ContextAwareLogger extends Logger {
+export class GlobalLogger extends Logger {
     constructor(loggerClients: LoggerClient[], currentProcess: typeof process) {
         super(loggerClients, currentProcess);
+    }
+
+    public async setup(baseProperties?: BaseTelemetryProperties): Promise<void> {
+        await super.setup(baseProperties);
     }
 }
