+++ packages/logger/src/index.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-export { Logger } from './logger';
+export { Logger, LogLevel } from './logger';
 export { ContextAwareLogger } from './context-aware-logger';
-export { LogLevel } from './base-logger';
 export { loggerTypes } from './logger-types';
 export { registerLoggerToContainer } from './register-logger-to-container';
 export { BaseTelemetryProperties } from './base-telemetry-properties';
