+++ packages/scan-request-sender/src/test-utilities/mockable-logger.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import { Logger } from 'logger';
+
+export class MockableLogger extends Logger {}
