+++ packages/service-library/src/register-service-library-to-container.spec.ts
@@ -7,17 +7,18 @@ import { Container } from 'inversify';
 import { Logger } from 'logger';
 import { IMock, Mock } from 'typemoq';
 import { registerServiceLibraryToContainer } from './register-service-library-to-container';
+import { MockableLogger } from './test-utilities/mockable-logger';
 import { WebDriver } from './web-driver/web-driver';
 
 // tslint:disable: no-unsafe-any no-any
 
 describe(registerServiceLibraryToContainer, () => {
     let container: Container;
-    let loggerMock: IMock<Logger>;
+    let loggerMock: IMock<MockableLogger>;
 
     beforeEach(() => {
         container = new Container({ autoBindInjectable: true });
-        loggerMock = Mock.ofType(Logger);
+        loggerMock = Mock.ofType(MockableLogger);
 
         container.bind(Logger).toConstantValue(loggerMock.object);
     });
