+++ packages/service-library/src/test-utilities/db-mock-helpers.ts
@@ -7,10 +7,10 @@ import { CosmosClientWrapper, CosmosOperationResponse } from 'azure-services';
 import { HashGenerator } from 'common';
 import * as crypto from 'crypto';
 import * as _ from 'lodash';
-import { Logger } from 'logger';
 import { ItemType, StorageDocument, Website, WebsitePage, WebsitePageExtra } from 'storage-documents';
 import { Mock } from 'typemoq';
 import { PageObjectFactory } from '../factories/page-object-factory';
+import { MockableLogger } from './mockable-logger';
 
 export interface DbContainer {
     dbName: string;
@@ -45,7 +45,7 @@ export class DbMockHelper {
         };
 
         this.azureCosmosClient = new cosmos.CosmosClient({ endpoint: cosmosDbUrl, auth: { masterKey: cosmosDbKey } });
-        this.cosmosClient = new CosmosClientWrapper(() => Promise.resolve(this.azureCosmosClient), Mock.ofType<Logger>().object);
+        this.cosmosClient = new CosmosClientWrapper(() => Promise.resolve(this.azureCosmosClient), Mock.ofType<MockableLogger>().object);
 
         await this.deleteDbContainer(this.dbContainer);
         await this.createDbContainer(this.dbContainer);
