+++ packages/web-api-scan-job-manager/src/batch/batch.spec.ts
@@ -7,9 +7,9 @@ import { BatchServiceClient, BatchServiceModels, Job, Pool, Task } from '@azure/
 import { JobGetTaskCountsResponse, JobListResponse, PoolGetResponse } from '@azure/batch/esm/models';
 import { Message, StorageContainerSASUrlProvider } from 'azure-services';
 import { ServiceConfiguration, TaskRuntimeConfig } from 'common';
-import { Logger } from 'logger';
 import * as moment from 'moment';
 import { IMock, It, Mock, Times } from 'typemoq';
+import { MockableLogger } from '../test-utilities/mockable-logger';
 import { BatchServiceClientProvider } from '../web-api-job-manager-ioc-types';
 import { Batch } from './batch';
 import { BatchConfig } from './batch-config';
@@ -41,7 +41,7 @@ describe(Batch, () => {
     let poolMock: IMock<Pool>;
     let storageContainerSASUrlProviderMock: IMock<StorageContainerSASUrlProvider>;
     let batchServiceClientProviderStub: BatchServiceClientProvider;
-    let loggerMock: IMock<Logger>;
+    let loggerMock: IMock<MockableLogger>;
     let taskEnvSettings: BatchServiceModels.EnvironmentSetting[];
     let taskResourceFiles: BatchServiceModels.ResourceFile[];
     let serviceConfigMock: IMock<ServiceConfiguration>;
@@ -79,7 +79,7 @@ describe(Batch, () => {
             pool: poolMock.object,
         } as unknown) as BatchServiceClient;
 
-        loggerMock = Mock.ofType(Logger);
+        loggerMock = Mock.ofType(MockableLogger);
         batchServiceClientProviderStub = async () => batchClientStub;
         storageContainerSASUrlProviderMock = Mock.ofType(StorageContainerSASUrlProvider);
         storageContainerSASUrlProviderMock
